Fix: 2021-06-14

+++ xsevenzip.cpp
@@ -490,14 +490,14 @@ qint64 XSevenZip::getUnpackInfo(qint64 nOffset, XSevenZip::XUNPACKINFO *pUnpackI
                 nCurrent+=pn.nByteSize;
 
                 // TODO Check!
-                for(int i=0;i<nFolderCount;i++)
+                for(int i=0;i<(qint32)nFolderCount;i++)
                 {
                     pn=get_packedNumber(nOffset+nCurrent);
 
                     quint64 nNumberOfCoders=pn.nValue;
                     nCurrent+=pn.nByteSize;
 
-                    for(int j=0;j<nNumberOfCoders;j++)
+                    for(int j=0;j<(qint32)nNumberOfCoders;j++)
                     {
                         // TODO read coder function
                         pn=get_packedNumber(nOffset+nCurrent);
@@ -513,7 +513,7 @@ qint64 XSevenZip::getUnpackInfo(qint64 nOffset, XSevenZip::XUNPACKINFO *pUnpackI
                 {
                     nCurrent+=pn.nByteSize;
 
-                    for(int i=0;i<nFolderCount;i++)
+                    for(int i=0;i<(qint32)nFolderCount;i++)
                     {
                         pn=get_packedNumber(nOffset+nCurrent);
